tydzien2/meeting-notes-wizard/requirements.txt
streamlit==1.50.0
openai==2.6.1
python-dotenv==1.1.1